// frontend/src/api/generated/models/createAiRegressionRunDto.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * 在线答题与智能测评平台 API
 * Online exam and intelligent assessment platform API
 * OpenAPI spec version: 1.0
 */
import type { CreateAiRegressionRunDtoSummaryType } from './createAiRegressionRunDtoSummaryType';

export interface CreateAiRegressionRunDto {
  /**
     * @minimum 1
     * @maximum 3
     */
  caseCount?: number;
  configId: string;
  summaryType: CreateAiRegressionRunDtoSummaryType;
}